16bit Z buffer and lores/hires switch

+++ examples/3d-demo/renderer.hpp
@@ -6,4 +6,4 @@ using namespace blit;
 
 extern uint32_t pixels_drawn;
 
-void draw_face(Vec3* vertices, Vec3* normals, Vec2* texture_coordinates, Surface* texture, Vec3 light, Pen* color, float *zbuffer, float near, float far);
+void draw_face(Vec3* vertices, Vec3* normals, Vec2* texture_coordinates, Surface* texture, Vec3 light, Pen* color, int16_t *zbuffer, int16_t near, int16_t far);
